% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thrust_changes.r
\name{thrust_changes}
\alias{thrust_changes}
\title{thrust_changes}
\usage{
thrust_changes()
}
\description{
See the list of file/folder name changes in the internal Thrust C++ library.
}
\examples{
thrust::thrust_changes()

}
